more refactoring #174
